+++ app/controllers/questions_controller.rb
@@ -8,7 +8,7 @@ skip_before_action :verify_authenticity_token, :only => :upload
     csv_que = CSV::parse(File.open('public/csv/questions.csv', 'r') {|f| f.read })
     fields = csv_que.shift
     @questions = csv_que.collect { |record| Hash[*fields.zip(record).flatten ] }
-    @questions = @questions.group_by { |d| d["qid"].to_i } 
+    @questions = @questions.group_by { |d| d["qid"].to_i }
     $user_data.select{|user| user[:computer_id] == "#{session[:computerid]}"}[0][:status]="Doing Quiz"
   end
