+++ app/views/home/index.html.erb
@@ -1,29 +1,34 @@
+<div class="container-fluid">
+	<div class="row">
 <h4 align="center">Welcome to Faces Procedures</h4>
 
 <%if @from=="exit"%>
-<h1 class="score" disabled="disabled">
+<div class="col-md-12">
+<h1 class="score text-center">
 	<%= "Your Earning is #{$userscore}"%>
 </h1>
+
 <%end%>
 
+<h1 class="thanks text-center">
+	<%= "Thank You!"%>
+</h1>
+</div>
 <div>
   <img src="<%= asset_path 'loader.gif' %>" class="loaderimg" />
 </div>
 
+</div>
+</div>
 
-<style type="text/css">
-  img.loaderimg{
-    display: block;
-    margin-left: auto;
-    margin-right: auto;
-    padding-top: 250px;
-  }
+<script src="http://code.jquery.com/jquery-1.11.0.min.js"></script>
+<script>
+$(document).ready(function(){
+	$('.thanks').hide();
+});
 
-  div.container {
-    min-height: 10em;
-    padding-left: 615px;
-    padding-top: 25px;
-    vertical-align: bottom;
-  }
-
-</style>
+	setTimeout(function() {
+		$('.score').hide();
+		$('.thanks').show();
+}, 10000);
+</script>
