+++ app/views/layouts/application.html.erb
@@ -2,21 +2,26 @@
 <html>
 <head>
   <title>PredictiveExperiment</title>
-  <div class="header">
-  <header>
-  	<h3 align="center">PredictiveExperiment</h3>
-  	</header>
+  <div align="right">
+  <%if session[:computerid]%>
+  <%="Round #{$round+1} / #{session[:computerid]}"%>
+  <%end%>
   </div>
-   <%= stylesheet_link_tag  'application',  media: 'all', 'data-turbolinks-track' => true %>
-      <%= javascript_include_tag 'application', 'recordrtc', 'data-turbolinks-track' => true %>
+  <%= stylesheet_link_tag  'application',  media: 'all', 'data-turbolinks-track' => true %>
   <%#= stylesheet_link_tag  'application', "bootstrap", "bootstrap.min", media: 'all', 'data-turbolinks-track' => true %>
   <%#= javascript_include_tag 'application', "jquery_v1.11.2", "bootstrap", "bootstrap.min", 'data-turbolinks-track' => true %>
   <%= csrf_meta_tags %>
 </head>
 <body>
 
-<%= yield %>
+<div class="header">
+  <header>
+    <h3 align="center">PredictiveExperiment</h3>
+    </header>
+  </div>
 
+<%= yield %>
+<%= javascript_include_tag 'application', 'recordrtc', 'data-turbolinks-track' => true %>
 </body>
 </html>
 <style>
