+++ app/views/participant/_participant.html.erb
@@ -1,23 +1,17 @@
-<div id="experiment" >
-
-
-  <% if @page == "waiting" %>
-    <%= render :partial => 'participant/waiting' %>
-  <% end %>
-
-
-  <% if $experiment_status != "start" and  @page == "adjust webcam"%>
-    <%= render :partial => 'participant/adjust_screen', :locals => { :from => "adjsut_cam" } %>
-  <% end %>
-  
-  <%if @page=="quiz"%>
-  <script type="text/javascript">
-    window.location.href="/quiz"  // put your correct path in a string here
-  </script>
-  <%elsif @page=="statement"%>
-  <script type="text/javascript">
-    window.location.href="/statements"  // put your correct path in a string here
-  </script>
-  <%end%>
-  
+<div id="waitmsgone">
+	<% if @page == "waiting" %>
+	<%= render :partial => 'participant/waiting' %>
+	<% end %>
+	<% if $experiment_status != "start" and @page == "adjust webcam"%>
+	<%= render :partial => 'participant/adjust_screen' %>
+	<% end %>
+	<%if @page=="quiz"%>
+	<script type="text/javascript">
+		window.location.href = "/quiz" // put your correct path in a string here
+	</script>
+	<%elsif @page=="statement"%>
+	<script type="text/javascript">
+		window.location.href = "/statements" // put your correct path in a string here
+	</script>
+	<%end%>
 </div>
