+++ app/views/participant/sample_video.html.erb
@@ -1,16 +1,20 @@
-<!DOCTYPE html>
-<html>
-  <head lang="en">
-    <meta charset="UTF-8">
-  </head>
-  <body>
+<div class="container-fluid">
 
 
+<%if @from!="result"%>
+<!--
 <div class="options">
   <button class="split" value="split" type="button">Split</button>
   <button class="takeall" value="takeall" type="button">Take All</button>
+</div>-->
+<div class="row">
+	<div class="col-md-6 col-md-offset-5" id="video-btn">
+		<a href="#" class="btn btn-success" id="split">Split</a> &nbsp;&nbsp;&nbsp;&nbsp;
+		<a href="#" class="btn btn-success" id="takeall">Take All</a>
+	</div>
 </div>
-    
+<%end%>
+    <%=hidden_field 'coming','from', :value=>@from%>
     <div id="views">
 
     </div>
@@ -27,7 +31,8 @@
 <div id="videodiv">
   <video id="video-playback" ></video>
 </div>
-
+</div>
+<script src="http://code.jquery.com/jquery-1.11.0.min.js"></script>
   <script>
 
     var counter;
@@ -79,7 +84,6 @@
     });
 
     function start() {
-    	alert("hid")
       navigator.getUserMedia(videoConstraints, startRecording, onfailure);
     }
 
@@ -184,6 +188,13 @@ $('button').click(function() {
         });
 	}
 });
+
+setTimeout(function() {
+    if ($('#coming_from').val() == "result"){
+        	window.location.href="/results?from=vdo";
+    }
+}, 10000);
+
 </script>
 
 </html>
