+++ public/csv/questions.csv
@@ -1,31 +1,32 @@
 qid,ref,data,answer
-1,0,This is the first question?,4
-1,1,This is not the answer x,0
-1,2,This is not the answer y,0
-1,3,This is not the answer z,0
-1,4,This is the answer,0
-2,0,This is the second question?,2
-2,1,This is not the answer x,0
-2,2,This is the answer,0
-2,3,This is not the answer y,0
-2,4,This is not the answer z,0
-3,0,This is the third question?,3
-3,1,This is not the answer x,0
-3,2,This is not the answer y,0
-3,3,Tihs is the answer,0
-3,4,This is not the answer z,0
-4,0,This is the Fourth question?,1
-4,1,Tihs is the answer,0
-4,2,This is not the answer x,0
-4,3,Tihs is not the answer y,0
-4,4,This is not the answer z,0
-5,0,This is the Fifth question?,1
-5,1,Tihs is the answer,0
-5,2,This is not the answer x,0
-5,3,Tihs is not the answer y,0
-5,4,This is not the answer z,0
-6,0,This is the Fifth question?,3
-6,1,This is not the answer x,0
-6,2,This is not the answer y,0
-6,3,Tihs is the answer,0
-6,4,This is not the answer z,0
+1,0,This is the first question that will be asked. It is to show that the question can be very long and it will display correctly.  Is this question display correctly?  Please choose your best response,3
+1,1,This is answer 1 that is not the correct answer.  If it were the correct answer then I would tell you,0
+1,2,This is answer 2 and again this is not the correct answer.  You need to move on.,0
+1,3,This is answer 3 and you guessed it this is the correct answer.  If you like you can read the last answer.,0
+1,4,This is answer 4 and why are you reading this since answer 3 was the correct answer.  I could give this a longer response to see how well a long response is displayed on the answer line.  Not sure how this will look but I figured I would ramble on a bit to keep this answer going ,0
+2,0,How does this look?,2
+2,1,Okay!,0
+2,2,Alright I guess!*,0
+2,3,Great!,0
+2,4,Couldn't be better!,0
+3,0,Incentives matter?,4
+3,1,only when people are greedy and selfish.,0
+3,2,only in a free market system.,0
+3,3,only in the private sector.,0
+3,4,to all human beings regardless of environment.,0
+4,0,Which of the following is often referred to as the basic postulate of economics?,2
+4,1,Individuals act only out of selfish motives.,0
+4,2,Incentives matterâ€”individuals respond in predictable ways to changes in personal costs and benefits.,0
+4,3,The accuracy of the assumptions is the best test of an economic theory.,0
+4,4,The value of a good is objective; it is equal to the cost of producing the good.,0
+5,0,3.Which of the following is a predictable secondary effect of a sharp increase in gasoline prices?,2
+5,1,Producers will increase the production of gas guzzling vehicles.,0
+5,2,Producers will increase the production of fuel-efficient cars.,0
+5,3,the termination of research on the cost-effectiveness of alternative fuels to power automobiles.,0
+5,4,The federal government will place a quota on the number of fuel-efficient cars for sale thus  consumers to purchase the gas guzzling vehicles.,0
+6,0,In a free market economy consumption and investment decisions,2
+6,1,are controlled largely by the government.,0
+6,2,shape the future course of the national economy.,0
+6,3,are necessarily controlled by big businesses,0
+6,4,require protection from foreign forces if individuals desire wealth accumulation.,0
+
